requirements.txt
curl_cffi
ddddocr
lxml
pycryptodome
selenium
